PathTracing/requirements.txt
numpy>=1.20.0
Pillow>=9.0.0